widget: Speed up the drawing of curly underline by caching its look
https://bugzilla.gnome.org/show_bug.cgi?id=721761

+++ src/vteinternal.hh
@@ -655,7 +655,7 @@ public:
         long m_double_underline_position;
         long m_double_underline_thickness;
         double m_undercurl_position;
-        long m_undercurl_thickness;
+        double m_undercurl_thickness;
         long m_strikethrough_position;
         long m_strikethrough_thickness;
         long m_overline_position;
